fix queries from direct space state(support for kinematic bodies). Also update to godot-cpp 4.1
Added implementation for:

fix _intersect_ray
fix _intersect_point
_intersect_shape
_cast_motion
_collide_shape
_rest_info

+++ src/shapes/box2d_shape_separation_ray.cpp
@@ -13,7 +13,7 @@ void Box2DShapeSeparationRay::set_data(const Variant &p_data) {
 	ERR_FAIL_COND(dict.size() != 2);
 	ERR_FAIL_COND(!dict.has("length"));
 	ERR_FAIL_COND(!dict.has("slide_on_slope"));
-	float length = dict["length"];
+	double length = dict["length"];
 	bool slide_on_slope = dict["slide_on_slope"];
 	a = Vector2();
 	b = Vector2(0, length);
